Add in authentication
Basically to sign in you have to create an instance of the Auth class, making sure to pass the activity to it (via 'this').

You can then call the login function, which has a callback that will give you the JWT token.

If the user has already logged in, then the JWT token will be passed automatically.

Else it will try to login the user automatically via Google (given they have a umich.edu account signed in.)